NS: fixed bug where NS failed to process Ack flag in US TX
Due to imporper initialization of the LoRaWANEndDeviceInfoNS structs
the value of the frame counter up was undefined in the struct. This
caused the NS to treat all US packets as a retransmission, thereby not
processing the Ack bit in the frame header which caused the NS to keep
on retransmitting the same packet before dropping it.